Move file to models folder
