+++ app/Events/Models/Users/UserUpdated.php
@@ -1,0 +1,34 @@
+<?php
+
+namespace App\Events\Models\Users;
+
+use App\Models\User;
+use Illuminate\Broadcasting\InteractsWithSockets;
+use Illuminate\Broadcasting\PrivateChannel;
+use Illuminate\Foundation\Events\Dispatchable;
+use Illuminate\Queue\SerializesModels;
+
+class UserUpdated
+{
+    use Dispatchable, InteractsWithSockets, SerializesModels;
+    protected User $user;
+    /**
+     * Create a new event instance.
+     */
+    public function __construct(User $user)
+    {
+        $this->user = $user;
+    }
+
+    /**
+     * Get the channels the event should broadcast on.
+     *
+     * @return array<int, \Illuminate\Broadcasting\Channel>
+     */
+    public function broadcastOn(): array
+    {
+        return [
+            new PrivateChannel('channel-name'),
+        ];
+    }
+}
