+++ app/Subscribers/Models/UserSubscriber.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace App\Subscribers\Models;
+
+class UserSubscriber
+{
+
+}
